@@ -15,7 +15,7 @@
     ],
     "minimum-stability": "dev",
     "require": {
-		"php": ">=5.0.0"
+		"php": ">=7.4.0"
 	},
     "autoload": {
         "psr-4": {
